requirements.txt
mysql-connector-python
mariadb
psycopg2
sqlite3